+++ htdocs/google/lib/google_calendar.lib.php
@@ -266,7 +266,9 @@ function createEvent($client, $object, $login='primary')
 
 	$event->setSummary(trim($object->label));
 	$event->setLocation($object->location);
-	$event->setDescription(dol_string_nohtmltag(($object->note_public ? $object->note_public : $object->note), 0));
+
+	$note = html_entity_decode(($object->note_public ? $object->note_public : $object->note), ENT_QUOTES);    // Because dol_string_nohtmltag does not convert simple quotes
+	$event->setDescription(dol_string_nohtmltag($note, 0));
 
 	// Transparency 0=available, 1=busy
 	$transparency=isset($object->userassigned[$user->id]['transparency'])?$object->userassigned[$user->id]['transparency']:0;
@@ -409,7 +411,9 @@ function updateEvent($client, $eventId, $object, $login='primary', $service=null
 
 		$event->setSummary(trim($object->label));
 		$event->setLocation($object->location);
-		$event->setDescription(dol_string_nohtmltag(($object->note_public ? $object->note_public : $object->note), 0));
+
+		$note = html_entity_decode(($object->note_public ? $object->note_public : $object->note), ENT_QUOTES);    // Because dol_string_nohtmltag does not convert simple quotes
+		$event->setDescription(dol_string_nohtmltag($note, 2));
 
 		// Transparency 0=available, 1=busy
 		$transparency=isset($object->userassigned[$user->id]['transparency'])?$object->userassigned[$user->id]['transparency']:0;